Add files via upload
In this project, we are going to examine how several factors, such as education, environment satisfaction and so on, affect the employee attrition in a company. This is an important issue for all companies, because hiring and retaining new employees is an extremely challenging task, requiring a lot of time and revenue. Hence, the existence of a detailed plan analyzing this problem is absolutely necessary. More particularly, we are going to analyze, which emplloyees are more likely to leave the company depending on several different factors. We are going to use an IBM fictional dataset for our analysis. Initially, we do our analysis by presenting several visualizations to draw conclusions on how the different factors affect the attrition of employees. Then, we build, train and validate several different classification algorithms. We also use the oversampling SMOTE method to mitigate the imbalanced nature of our dataset. Finally, we compare all classification models for both the original and transformed datasets.